requirements.txt
Flask==2.0.1
SQLalchemy==1.4.17
Flask-SQLalchemy==2.5.1
Flask-Migrate==3.0.1
Flask-Login==0.6.2
Werkzeug==2.3.4
gunicorn==21.2.0
python-dotenv==0.19.1